# Vakaros.Vkx.Api/OpenApi/Vakaros.Vkx.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GetDocument.Insider | v1",
    "description": "GetDocument Command-line Tool inside man",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/Boats/{id}": {
      "get": {
        "tags": [
          "Boats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BoatDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoatDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoatDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Boats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BoatDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoatDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoatDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Boats"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Boats/{id}/stats": {
      "get": {
        "tags": [
          "Boats"
        ],
        "summary": "Aggregate statistics for a boat across all its sessions and races.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BoatStatsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoatStatsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoatStatsDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/Courses/{id}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoatClassSummaryDto": {
        "required": [
          "id",
          "name",
          "length",
          "width",
          "weight"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "length": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "width": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "weight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "BoatDto": {
        "required": [
          "id",
          "name",
          "sailNumber",
          "boatClass",
          "description",
          "isPublic",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "sailNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "boatClass": {
            "$ref": "#/components/schemas/BoatClassSummaryDto"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPublic": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BoatStatsDto": {
        "required": [
          "boatId",
          "boatName",
          "sailNumber",
          "boatClass",
          "sessionCount",
          "raceCount",
          "totalSessionDurationSeconds",
          "totalRaceDurationSeconds",
          "totalSailedDistanceMeters",
          "topSpeedOverGround"
        ],
        "type": "object",
        "properties": {
          "boatId": {
            "type": "string",
            "format": "uuid"
          },
          "boatName": {
            "type": "string"
          },
          "sailNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "boatClass": {
            "$ref": "#/components/schemas/BoatClassSummaryDto"
          },
          "sessionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "raceCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalSessionDurationSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalRaceDurationSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalSailedDistanceMeters": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "topSpeedOverGround": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "float"
          }
        }
      },
...
      "CourseDto": {
        "required": [
          "id",
          "name",
          "year",
          "description",
          "createdAt",
          "totalLengthMeters",
          "startLineSource",
          "startMark1Id",
          "startMark2Id",
          "finishLineSource",
          "finishMark1Id",
          "finishMark2Id",
          "legs"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "totalLengthMeters": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "startLineSource": {
            "type": "string"
          },
          "startMark1Id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "startMark2Id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "finishLineSource": {
            "type": "string"
          },
          "finishMark1Id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "finishMark2Id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "legs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseLegDto"
            }
          }
        }
      },
      "CourseLegDto": {
        "required": [
          "id",
          "markId",
          "markName",
          "gateMarkId",
          "gateMarkName",
          "sortOrder",
          "legName",
          "legType",
          "passingSide",
          "latitude",
          "longitude",
          "gateLatitude",
          "gateLongitude"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "markId": {
            "type": "string",
            "format": "uuid"
          },
          "markName": {
            "type": "string"
          },
          "gateMarkId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "gateMarkName": {
            "type": [
              "null",
              "string"
            ]
          },
          "sortOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "legName": {
            "type": [
              "null",
              "string"
            ]
          },
          "legType": {
            "type": "string"
          },
          "passingSide": {
            "type": "string"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "gateLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "gateLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "CourseLegRequest": {
        "required": [
          "markId",
          "gateMarkId",
          "legName",
          "legType",
          "passingSide"
        ],
        "type": "object",
        "properties": {
          "markId": {
            "type": "string",
            "format": "uuid"
          },
          "gateMarkId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "legName": {
            "type": [
              "null",
              "string"
            ]
          },
          "legType": {
            "type": "string"
          },
          "passingSide": {
            "type": "string"
          }
        }
      },
...
      "UpdateBoatRequest": {
        "required": [
          "name",
          "sailNumber",
          "boatClassId",
          "description",
          "isPublic"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "sailNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "boatClassId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPublic": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "UpdateCourseRequest": {
        "required": [
          "name",
          "year",
          "description",
          "startLineSource",
          "startMark1Id",
          "startMark2Id",
          "finishLineSource",
          "finishMark1Id",
          "finishMark2Id",
          "legs"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "startLineSource": {
            "type": [
              "null",
              "string"
            ]
          },
          "startMark1Id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "startMark2Id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "finishLineSource": {
            "type": [
              "null",
              "string"
            ]
          },
          "finishMark1Id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "finishMark2Id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "legs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseLegRequest"
            }
          }
        }
      },
...
    }
  },
...
}