# Vakaros.Vkx.Shared/Dtos/Boats/BoatDto.cs
using Vakaros.Vkx.Shared.Dtos.BoatClasses;

namespace Vakaros.Vkx.Shared.Dtos.Boats;

public record BoatDto(Guid Id,
                      string Name,
                      string? SailNumber,
                      BoatClassSummaryDto BoatClass,
                      string? Description,
                      bool IsPublic,
                      DateTimeOffset CreatedAt);